persist: introduce filter API boundary between persist+compute
Persist is introducing the ability to skip fetching a chunk of data
("part") at read time if we can tell from statistics about it (e.g.
min/max) that it would entirely be filtered out by an MFP once fetched.

This is a performance optimization, but it is believed it will be enough
to unlock acceptable rehydration times for temporal filters on columns
of data that are roughly correlated with realtime. It may also help with
one-off exploratory queries. See #18007 for more context.

The primary purpose of this PR is to introduce an initial guess at the
API boundary between persist/storage (which contains the stats and
fetching) and compute (which contains MFP and expression evaluation) for
these filters. In particular, if persist_source is given an MFP, then
persist asks the MFP a series of queries about whether it can prove that
no data in a part is relevant, given some stats.

For crate dependency and code ownership reasons, the pieces are broken
up as follows:
- In mz_repr, a `PersistSourceDataStats` trait which can answer
  questions about which `SourceData` (`Result<Row, DataflowError>`) may
  be present in a part. This is a trait for dependency reasons, but it
  should also make it easier to write unit tests.
- In mz_expr next to MapFilterProject/MfpPlan, an `MfpPushdown`
  implementation of the query logic mentioned above.
- In mz_storage_client next to persist_source, an implementation of the
  `PersistSourceDataStats` trait that wraps persist's internal stats
  format and maps it back to Datums, etc.

This PR is the last piece necessary to getting everything working
end-to-end (for a _very_ limited subset of queries) behind feature
flags. As a result, it also includes a placeholder impl of the
MfpPushdown logic. This is not meant to be representative of what the
final logic will look like, but rather will allow us to start getting
this in front of devex (customers?) ASAP so that we can derisk the
feature itself as well as the initial scope.