sql: introduce dataflow error if PG data violates NOT NULL constraint
The PG replication stream does include changes to column's constraints,
and we will ultimately need to rely on querying to upstream PG
database to determine if these values have changed at all. Even
in the case where we get a relation message regarding these changes,
it's possible our understanding of the NOT NULL constraint will be
out of date, and we will ingest NULL datums. The appropriate time
to error if on ingesting the data without letting it propagate to
any downstream dataflows, so manufacture that error here.