adapter: consistent global introspection sources
This commit refactors existing introspection sources and views, and
introduces new ones, to expose per-worker and global introspection
relations in a consistent manner.

Introspection sources and views that are per-worker (i.e. they have a
`worker_id` column) are suffixed with `_per_worker`. Exempt from this
rule are `*_raw` sources. These are always per-worker and I found it too
noisy to suffix them all with `_per_worker_raw`.

Each `*_per_worker` relation has a corresponding global view defined on
it that omits the suffix and the `worker_id` column. The remaining
columns are aggregated in a sensible way, depending on the semantics of
the relation. For example, relations reporting dataflow structure simply
take the values from worker 0, whereas histogram relations compute the
sum of the per-worker counts.

We expect that the global view will be more useful to users than the
per-worker ones, which is why we add a suffix to the latter rather than
the former.