storage: fix correctness issue of load generators
The load generator sources had a correctnes issue where upon restart
they would fast foward exactly as many rows as the offset of their
resumption frontier.

This only works under the assumption that each offset contains exactly
one row which is true for the counter generator but not true for the
TPCH generator.

The fix is to simply count the number of transactions when fast
fowarding the iterator.

Fixes #16259

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>