adapter: move declare impl into background thread
Allow the declare impl (which does non-trivial work for name resolution
and result column type descriptions) to run in a background thread with
a catalog snapshot, allowing the coordinator to continue working. This
won't affect internal correctness because declare records the catalog
transient revision, so even if the catalog has changed before this
function completes, it has recorded the revision at which it was
created. This doesn't violate SS because two statements can execute
concurrently in any order.