Remove persisted introspection sources
This commit removes the persisted introspection sources from the
codebase. These have been disabled for a long time due to the load they
place on persist. In the meantime we have found that the arranged
introspection sources serve as a usable alternative for our debugging
purposes.

Even though persist has meanwhile improved its performance to the point
where it can almost support the workload caused by the persisted
introspection sources, those sources still don't provide the ergonomics
we would like. They require users to pick a per-replica relation by
suffixing it with the replica ID, so they are not more convenient to use
than the arranged introspection sources. They also litter `mz_sources`
with additional relations for every existing replica, which causes a lot
of noise. Their only advantage is that they make it possible to query
introspection data from a replica that is crashing fast. So far we
haven't had much need for this feature though.

We are still looking for ways to make introspection ergonomic. The final
solution may or may not use persist. If it uses persist, it will
likely look different than the current implementation and therefore
would require a substantial rewrite anyway. Based on this, there is not
much reason to keep the current implementation around. By removing it,
we remove the burden of maintaining its code, which it quite convoluted
at times.