adapter: Prevent statements with dropped roles (#18334)
Previously, if a user's role was dropped while they were connected,
then we'd prevent them from executing a statement only if RBAC checks
were turned on. This commit modifies the logic, so we always prevent
dropped roles from executing a statement even if RBAC checks are turned
off.

We will need this check for object ownership. If a user's role was
dropped, and then they try and create an object, we will have no role
to assign as the object's owner.

Part of #11579