mzcompose: use multiple concurrent profraw files to avoid collisions
The goal here is to avoid sqllogictest and clusterd from overwriting
each-other's profraw outputs.  %p is by PID, which one would expect to
be sufficient, and %Nm allows up to 9 concurrent profiles for the same
build-ID; perhaps the two together is overkill, but this seems like a
reasonable starting point.