storage: cleanup leaked IDs on boot
It's possible for the adapter to forget about IDs that the storage
controller remembers in COLLECTION_METADATA; in these cases, we
will permanently leak the shards associated with the ID if they
are not finalized. To avoid doing this work in perpetuity on reboot,
also remove the ID from COLLECTION_METADATA after durably recording
the shards to finalize.