doc: fix link in developer readme
The "Architecture Overview" link in the developer
README points to a non-existent page in the public
docs. There's an open PR to add an architecture
page to the docs, but it's not clear when that will
merge. In the meantime, this PR fixes the 404 by
linking to the Materialize Platform architecture
doc in the repo.

Fixes #16474

Open PR mentioned above: #16141