persist: remove some SpineBatch::clone calls
These also seemed to be showing up in profiling in #18368.

Accomplished by removing the `SpineBatch`s from `FuelingMerge` and
instead handing them back in the `done` call. This has the side benefit
of cleaning up some complexity in `apply_merge_res` where we had to make
sure we applied the merge results to both copies. I didn't do this
originally because it further deviates from the DD impl of Spine, but
the perf seems worth it.