storage: convert kafka to `SourceRender`
This commit applies a few simplifications in the process of converting
the source to the `SourceRender` trait by being smarter with how
capabilities are managed.

Specifically, the operator does *not* downgrade its capability for every
message. Instead, it exhausts all the available messages from the
partitions which are emitted with the same capability but at different
times (a la DD triplets) and then downgrades the capability of each
partition onces.

This produces a significant performance boost in the feature benchmarks:

```
NAME                      |    THIS     |    OTHER    |  Regression?  | 'THIS' is:
----------------------------------------------------------------------------------------------------
KafkaEnvelopeNoneBytes    |      11.524 |      12.409 |      no       | 7.1 pct   faster
KafkaUpsert               |       3.163 |       5.172 |      no       | 38.8 pct   faster
KafkaUpsertUnique         |       4.436 |       5.180 |      no       | 14.4 pct   faster
```

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>