Fix monotonic top-k bug #18445
This commit fixes a bug (#18445) in monotonic top-k evaluation. The intra-timestamp thinning step included code in TopKBatch that assumed that the value of K would always be smaller or equal to the number of updates in such a batch. This assumption is not true in general, thus creating incorrect output when it is not verified.