adapter: correct stash retraction migrations
When we add an Option field to an existing catalog struct, that gets
deserialized from the on-disk JSON as None. If we write a migration that
migrates all Nones to a Some, the storage transaction notices we changed
something so it issues a retraction. However the row it's retracting has
a "field: None" entry, but the row on disk has *no* entry, so the
retraction doesn't have a matching row from JSON's point of view, thus
never gets consolidated.

This oddly doesn't affect correctness because the Rust consolidation
code knows these cancel each other out, because they both deserialize
the same.

It does however affect us when we then remove the Optionness of the new
field, because we believe everything on disk is now populated. When we
try to read the row without the field at all into a Rust struct, the
migration errors.

To fix: write a transaction method that is able to tell apart things
that have JSON differences but are Rust-equivalent, and retract all the
rows (additions or retractions) that need to be consolidated from disk.
Use this function to cleanup all collections before running the stash
migration (a pre-migration migration), which can be removed after one
release. Prevent this from happening in the future by automatically
running this in all migrations for any changed collection.