upsert: dedup correctly, operate on (hash(key), value) pairs
This PR fixes a logic but in the upsert operator where it was previously
sorting updates based on `(ts, order)` and then deduplicating based on
`(ts, key)`. Of course this breaks completely if the topic contains
updates like:

```
(ts1, key1, order1)
(ts1, key2, order2)
(ts1, key1, order3)
(ts1, key2, order4)
...
```

because we never dedup anything.

This PR fixes that by sorting based on `(ts, key)`. During this work I
also took the opportunity to build the upsert based on `(hash(key),
value)` pairs instead of relying on trickly implementations of `Hash`
traits and ab-using a `HashSet`. This new implementation is going to be
compatible with the upcoming work of storing the hashmap out of core
where we can't rely on rust trait tricks.

Fixes #18466

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>