repr: hook up real columnar encodings for more ScalarTypes
For all of these, the columnar encoding's ordering matches the Datum's
ordering, so they are ones we could easily keep stats for.

To make this less boilerplate-y, we introduce the `enum_dispatch` crate.

Benchmarks seem to be slightly improved. Not sure why, but I've flipped
back and forth a bit and I think both are real.

    roundtrip_encode_structured
                            time:   [736.94 µs 737.33 µs 737.78 µs]
                            change: [-18.475% -18.351% -18.230%] (p = 0.00 < 0.05)
                            Performance has improved.

    roundtrip_decode_structured
                            time:   [633.84 µs 634.18 µs 634.57 µs]
                            change: [-3.8598% -3.7183% -3.5883%] (p = 0.00 < 0.05)
                            Performance has improved.