transform: use `base_eq` in `DatumKnowledge::join_assign`
Change the `DatumKnowledge::join` case where both sides are literals to
have higher overlap with the `ColumnType::union` logic, more
specifically in when `lhs.base_eq(rhs)`.

Relax the `soft_panic_or_log` call in the same place in `join` to a
`::tracing::error` call. The motivation is as follows:

- We don't want to panic in `join` when soft assertions are allowed
  because the old `ColumnKnowledge` version was also not handling
  these, so they should not be considered regressions.
- We still want to panic in the meet case because this is an indicator
  that the underlying tree is not well-typed.