persist: collect schemaless stats for Jsonb
Do this by allowing Schema impls to supply arbitrary logic for computing
stats.

It's a little gross, but to take advantage of this, we change the
associated `Stats` type of the `Data` impl for `Vec<u8>` to be a new
`BytesStats` enum. It has one variant for normal SQL bytes columns and a
Json variant for the new stats.

Given the new stats_fn override, use it in `TodoDatumToPersist` to
generate real mins and maxes by delegating the Datum's `Ord` impl.