storage: don't panic on source reader init errors
Instead, log the error to the source status collection and request a
`SuspendAndRestart`.

This commit also introduces a mandatory 30s backoff before a
`SuspendAndResume` takes place, to limit hot retry loops. This comes at
the expense of latency in the case of a transient error; unfortunately
implementing a smarter retry policy (like exponential backoff) is
difficult.

Fix #18491.