kafka-util: simplify BrokerRewritingClientContext
Rather than having separate APIs for rewrites that need to keep a token
alive and rewrites that don't, unify the API around closures. A rewrite
is now a closure that returns the rewritten URL. This allows the
rewrites that need to keep a token alive to do so by embedding the token
in the closure, with no special support required by the
BrokerRewritingClientContext.

A future commit will use this support to allow the rewrite addresses to
change over time to reflect a failed and restarted SSH tunnel.