storage,ccsr: automatically restart failed SSH tunnels
This commit reworks SSH tunnel handling so that tunnels are monitored
for failures and automatically restarted in the background, until all
outstanding handles to the tunnel are dropped.

This is expected to make a large improvement in the reliability of
SSH-tunneled Kafka sources in production. Unlike PostgreSQL sources,
which are restarted on connection failure, Kafka sources that experience
a connection failure are *not* restarted, as the Kafka consumer
internally retries the failed connection. But in the case of an
SSH-tunneled Kafka source, if the SSH tunnel failed, it would never be
recreated, and the Kafka consumer would attempt to reconnect to a broken
SSH tunnel forever.

The implementation is not trivial, as failed SSH tunnels must be
recreated with a new address, as the original local port may no longer
be available. To support this, tunneling a CSR client through SSH now
uses a reqwest::Proxy, rather than simply overriding the URL at client
creation time, to allow the tunnel URL to be dynamically updated over
time. This winds up being a net reduction in complexity, though, as the
SSH tunnel can be stored in the proxy closure, and so we no longer need
a special wrapper type to keep the SSH tunnel alive as long as the CSR
client is alive.