ore: add `ErrorExt` with `display_with_causes` and `to_string_with_causes`
These can print an error along with its chain of source errors, similar
to how `anyhow` can print the chain of errors when using the alternate
selector.

Using these will make it more explicit when/if we want to print the full
chain of errors instead of relying on `anyhow::Error` and it's printing
of the chain when selecting the alternate.