adapter: Fix mz_sessions retractions (#18611)
One of the columns in the `mz_sessions` system table is the time at
which the session connected. We were generating this timestamp at the
same time that we generated the update to the table. As a consequence
the addition update and the retraction update would have different
timestamps causing invalid retractions.

This commit fixes this issue by saving the timestamp as metadata with
each connection and passing in the timestamp to the row update
function. It also switches to using deterministic timestamps instead
of hard-coding in the system clock.

Fixes #18607