adapter: handle publications specified as identifiers
The parser currently accepts specifying the publication for a postgres
source with both single and double quotes which are parsed as strings
and identifiers respectively.

The catalog migration patched in this PR was making the assumption that
only strings are accepted, leading to a crash during environmentd
bootstrap when the catalog contained publication specified as idents.

Independently of this PR we should think about whether or not we want to
make a technically breaking change and disallow idents completely from
parser positions that expect a string.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>