adapter: refactor peek
This is only code movement. Add a new enum that describes each part of a
peek in order, so that it is really clear how they work and what each
part needs. We are then able to more clearly name the functions that process each part.

The second major benefit is this prepares us to split off some parts
into other threads, and it'll be obvious how to maintain clarity (i.e.,
add a new stage variant).