persist: bound the size of the seqno hold bug info! log
This was previously printing out the entire contents of the map, which
can potentially be large enough to overflow buffers when printed (as
recently seen by Philip). We don't get much value past seeing the
smallest few entries in it, so put a small bound on the number we keep.