persist: truncate String and Vec<u8> stats to a maximum byte length
These stats are inclusive lower and upper bounds, so we can (carefully)
truncate almost any String or Vec<u8> value to produce one that's
smaller but still a relatively tight bound.

There is TONS of stats pruning left to do (including given the
ProtoDatum and Json stats the same treatment), but this is enough to get
CI (which sometimes inserts 100MiB strings) to pass with the stats
collection ff on, so it seems like a nice start.