adapter: Refactor cluster replica dependents (#18693)
Previously, the assumption that cluster replicas had no dependents was
sprinkled throughout the code in multiple places. This commit
centralizes that assumption to the `cluster_replica_dependents` method.

This assumption hasn't changed, but moving it to a single place makes
the assumption more future-proof and easier to change, should it change
in the future.