sql: Add more variables for controlling RBAC (#18650)
This commit adds two additional system/session parameters to turn RBAC
on and off. The parameters that exist are as follows:

- A system parameter called "enable_ld_rbac_checks". This parameter will
only be able to be set by Launch Darkly, no users can toggle this
parameter. When this is off, then no user can enable RBAC. When this is
on, then a user can enable RBAC through one of the other parameters.
This acts as a global off switch for RBAC in case we ever need to ensure
that RBAC is off for a customer.
- A system parameter called "enable_rbac_checks". This parameter
can be set by any superuser. When this is on, then RBAC is enabled for
all users, as long as "enable_ld_rbac_checks" is also on.
- A session parameter called "enable_session_rbac_checks". This
parameter is session local and can be set by any user. When this is on,
then RBAC is enabled for the current session, as long as
"enable_ld_rbac_checks" is also on. The intention is that this can be
used to test RBAC in a single session, without having to turn it on for
everyone.

Part of #11579