Design document for invalid accumulations in reductions (#17597)
This commit adds a design document proposing strategies to handle invalid
accumulations in reductions. The two core strategies are to: (a) use of the Result type at the first level of a multi-level reduce operator hierarchy where the output arrangement of the first reduction is not reused downstream; (b) use a pair of reduce operators to redundantly process the input arrangement and compute either output or errors.