Factor common bucketing logic out of Reduce and TopK
This makes this logic consistent between the two, and will make it
easier for us to vary the fan-in strategies per #17135.

It seems reasonable to give TopK a default expected group size
matching reduce's default; if the user truly requires the impressive
tower of stages originally constructed, they can supply EXPECTED GROUP
SIZE = 18446744073709551615.