Use u64 consistently for EXPECTED GROUP SIZE
It's unclear why Reduce originally took this as a usize, but it must
be a 64-bit value, and we use it for determining bucketing that has
nothing to do with the word-size of the machine mz is running on.

This doesn't change anything in practice, except that on 32-bit
machines, I think the code wouldn't have even built, since `cast_from`
lacks an implementation for usize from u64.