Change schema of `mz_dataflow_channel_operators` (#18747)
The current view uses inner joins, so it does not contain channels that
don't correspond to a live operator. This can occur if an operator
downgrades its capabilities to empty and is therefore cleaned up by the
system. Much more commonly, it can happen with channels that enter or
leave a scope. By convention, these are considered to come from or to
slot 0 of the scope, which has no corresponding operator.

In this PR, we add the addresses of the channels' endpoints, so that we
can uniquely identify operators that have gone away and also enter/leave
channels.