persist: tune "snapshot ... not yet available" info log
The main questions this log answers are "was something held up because
an upper wasn't caught up" and "is such a case still ongoing", which
means we can reduce this log line from being printed N times to twice.

While I'm in here, also increase the threshold for switching from debug
to info, because this is semi-expected for small intervals when joining
sources to tables. Also remove a bit of noise from the line itself by
debug printing the Antichain elements.