frontegg-auth: compare emails case insensitively
Frontegg performs case-insensitive comparisons on emails. This commit
teaches Materialize to do the same.

At least one user has been tripped up by the old behavior of
case-sensitive comparisons; specifically, they were invited to their
Materialize Cloud account with an email spelled

    First.Last@corp.com

but were attempting to log in as first.last@corp.com. This worked to log
into the web UI, but not via pgwire.

The actual behavior change is in frontegg-auth and is quite
straightforward, but the diff has a lot of noise because:

  1. The auth tests required adaptation to distinguish between the user
     supplied to the pgwire/HTTP protocol (possibly wrongly cased) and
     the user reported by `SELECT current_user` (always canonically
     cased).

  2. The pgwire package was creating sessions using the username
     supplied to the protocol (possibly wrongly cased), rather than the
     user returned by Frontegg (always canonically cased).

  3. The `frontegg-auth` crate required substantial refactoring to avoid
     duplicating work between the initial claim validation and the
     continuous revalidation future.

User names inside of Materialize remain case sensitive. Users can type

    CREATE ROLE "first.last@corp.com";
    CREATE ROLE "First.Last@corp.com";

in order to create two distinct roles; however they will only be able to
log into whichever role has the casing that Frontegg has deemed
canonical for that email (i.e., whatever casing was used whenever the
user was invited to the tenant).