storage: bounded memory upsert snapshot
This makes recovering the upsert snapshot use memory proportinal to the
number of keys as opposed to the current method that uses memory
proportional to the total number of updates.

We use a XOR trick in order to accumulate the snapshot without having to
store the full unconsolidated history in memory. For all (value, diff)
updates of a key we track:
- diff_sum = SUM(diff)
- value_accum = XOR(bincode(value))
- checksum_accum = XOR(checksum(bincode(value)))

The method is correct because a well formed upsert snapshot will have
for each key:
- Zero or one updates of the form (cur_value, +1)
- Zero or more pairs of updates of the form (prev_value, +1), (prev_value, -1)

We are interested in extracting the cur_value of each key and discard
all prev_values that might be included in the stream. Since the history
of prev_values always comes in pairs, computing the XOR of those is
always going to cancel their effects out. Also, since XOR is commutative
this property is true independent of the order.

Therefore the accumulators will end up precisely in one of two states:
1. diff == 0, checksum == 0, value == [0..] => the key is not present
2. diff == 1, checksum == checksum(cur_value) value == cur_value => the
   key is present

In the absense of bugs, accumulating the diff and checksum is not
required since we know that a well formed snapshot always satisfies
XOR(bincode(values)) == bincode(cur_value). However bugs may happen and
so storing 16 more bytes per key to have a very high guarantee that
we're not decoding garbage is more than worth it.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>